move 'mapObjIndexed' tests to separate file
